Rename entity folder to model
