+++ src/main/java/com/yi/online_bookshop/domain/model/account/AccountFactory.java
@@ -1,6 +1,6 @@
-package com.yi.online_bookshop.domain.entity.account;
+package com.yi.online_bookshop.domain.model.account;
 
-import com.yi.online_bookshop.domain.entity.user.User;
+import com.yi.online_bookshop.domain.model.user.User;
 
 public class AccountFactory {
