+++ src/main/java/com/yi/online_bookshop/domain/model/payment/Payment.java
@@ -1,9 +1,9 @@
-package com.yi.online_bookshop.domain.entity.payment;
+package com.yi.online_bookshop.domain.model.payment;
 
 import java.math.BigDecimal;
 import java.time.LocalDate;
 
-import com.yi.online_bookshop.domain.entity.account.Account;
+import com.yi.online_bookshop.domain.model.account.Account;
 import com.yi.online_bookshop.enums.PaymentType;
 
 import jakarta.persistence.Column;
