+++ src/main/java/com/yi/online_bookshop/repository/AccountRepository.java
@@ -3,7 +3,7 @@ package com.yi.online_bookshop.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.yi.online_bookshop.domain.entity.account.Account;
+import com.yi.online_bookshop.domain.model.account.Account;
 
 @Repository
 public interface AccountRepository extends JpaRepository<Account, Long> {
