+++ src/main/java/com/yi/online_bookshop/repository/UserRepository.java
@@ -3,8 +3,7 @@ package com.yi.online_bookshop.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.yi.online_bookshop.domain.entity.user.User;
-
+import com.yi.online_bookshop.domain.model.user.User;
 
 @Repository
 public interface UserRepository extends JpaRepository<User, Long> {
